Bad input tests and refactoring tests
