+++ test/helper.js
@@ -3,6 +3,8 @@ let debug           = require('debug')('test:helper');
 let load            = require('../src/main');
 let config          = require('typed-env-config');
 let testing         = require('taskcluster-lib-testing');
+let urlencode       = require('urlencode');
+let EventSource = require('eventsource');
 
 const profile = 'test';
 let loadOptions = {profile, process: 'test'};
@@ -24,6 +26,20 @@ mocha.before(async () => {
   webServer = await load('server', loadOptions);
   debug('Server Setup');
 });
+
+helper.connect = bindings => {
+  let json = urlencode(JSON.stringify(bindings));
+  var es = new EventSource('http://localhost:12345/api/events/v1/connect/?bindings='+json);
+
+  var pass, fail;
+  var resolve = new Promise((resolve, reject) => {pass = resolve; fail= reject;});
+  return {
+    es:      es,
+    resolve: resolve,
+    pass:    pass,
+    fail:    fail, 
+  };
+};
   
 // Cleanup after tests
 mocha.after(async () => {
